[update] the LambdaRunnable to sit inside of root of the runnables module.
